Move header include path to rust/cxx.h
