+++ build.rs
@@ -1,8 +1,8 @@
 fn main() {
     cc::Build::new()
-        .file("src/cxxbridge.cc")
+        .file("src/cxx.cc")
         .flag("-std=c++11")
         .compile("cxxbridge01");
-    println!("cargo:rerun-if-changed=src/cxxbridge.cc");
-    println!("cargo:rerun-if-changed=include/cxxbridge.h");
+    println!("cargo:rerun-if-changed=src/cxx.cc");
+    println!("cargo:rerun-if-changed=include/cxx.h");
 }
