+++ src/cxx.cc
@@ -1,4 +1,4 @@
-#include "../include/cxxbridge.h"
+#include "../include/cxx.h"
 #include <cstring>
 #include <iostream>
 #include <memory>
